+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add array of transferable objects to the `execute()` method arguments.
+
 ## [2.6.23] - 2023-08-11
 
 ### Fixed